Update js/jquery.codebomber.panel.js
I had a need for loading static content within the panel not 
loading ajax content.  I added an option for static to reflect
if the user wanted to load static content.  Default is false, which
does not change behavior.  If static is set to true, the panel
is expanded and does not run base.load().